pyproject.toml
[project]
name = "notely"
version = "0.4.0"
dependencies = [
    "torch",
    "torchvision",
    "torchaudio",
    "openai-whisper",
    "ollama",
    "numpy",
    "halo"
]
authors = [
    {name = "TheWorldJar", email = "theworldjar@protonmail.com"}
]
description = "Create transcripts from audio files and process thoses transcripts to create notes!"
requires-python = ">= 3.9"
[project.optional-dependencies]
dev = [
    "pytest"
]
    [tool.black]
    line-length = 88
    target-version = ["py39"]
    include = '\.pyi?$'
    exclude = '''
        /(
          \.git
          | \.venv
          | build
          | dist
          | db
          | transcribe
          | note
        )/
    '''